data-engineering-curriculum/rag/agents-towards-production/tutorials/agent-with-streamlit-ui/requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
openai>=1.0.0
python-dotenv>=1.0.0 